Align release help with target catalogue (#551)
Expose the Clap command surface through release-help metadata so the
manual and PowerShell artefacts document `help targets`.

Harden target-help diagnostics against manifest control characters and
exercise root, topic, and no-Ninja command boundaries. Split the
catalogue integration scenarios into a focused child module.
